Dissolves app into two simple components Player and PlayList
